+++ changelog.md
@@ -21,6 +21,7 @@ This patch adds a number of new, useful features rather than simply fixing the b
 - Efficient Filter: much lower RAM usage + faster filtering by histogram selection (1D/2D), especially for very large datasets
 - Fixed Gauss-fitting error when spot's sum is zero (zero division error)
 - Fixed NND plot reindexing after fitting (#665)
+- Affine transform calibration for astigmatic imaging
 
 ## 0.10.0